+++ mex/extractors/international_projects/main.py
@@ -24,6 +24,7 @@ from mex.extractors.international_projects.models.source import (
 from mex.extractors.international_projects.transform import (
     transform_international_projects_sources_to_extracted_activities,
 )
+from mex.extractors.mapping.extract import extract_mapping_data
 from mex.extractors.pipeline import asset, run_job_in_process
 from mex.extractors.settings import Settings
 from mex.extractors.sinks import load
@@ -117,9 +118,16 @@ def extracted_international_projects_activities(
     ],
 ) -> list[ExtractedActivity]:
     """Transform projects to extracted activities, load and return them."""
+    settings = Settings.get()
+    activity = extract_mapping_data(
+        settings.international_projects.mapping_path / "activity.yaml",
+        ExtractedActivity,
+    )
+
     mex_sources = list(
         transform_international_projects_sources_to_extracted_activities(
             international_projects_sources,
+            activity,
             extracted_primary_source_international_projects,
             international_projects_person_ids_by_query,
             unit_stable_target_ids_by_synonym,