+++ mex/extractors/international_projects/settings.py
@@ -16,3 +16,10 @@ class InternationalProjectsSettings(BaseModel):
             "absolute path or relative to `assets_dir`."
         ),
     )
+    mapping_path: AssetsPath = Field(
+        AssetsPath("mappings/__final__/international-projects"),
+        description=(
+            "Path to the directory with the international-projects mapping files "
+            "containing the default values, absolute path or relative to `assets_dir`."
+        ),
+    )