+++ tests/seq_repo/test_extract.py
@@ -5,7 +5,7 @@ import requests
 from pytest import MonkeyPatch
 
 from mex.common.exceptions import MExError
-from mex.drop import DropApiConnector
+from mex.extractors.drop import DropApiConnector
 from mex.extractors.seq_repo.extract import extract_sources